api_com_pgsql/requirements.txt
alembic==1.6.5
aniso8601==9.0.1
apispec==4.4.2
autopep8==1.5.7
blinker==1.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Email==1.4.4
Flask-JWT-Extended==4.1.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.0.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
passlib==1.7.4
Pillow==8.3.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.12
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
webargs==8.0.0
Werkzeug==1.0.1
zipp==3.4.1
